Change root dynamic property to false and make prefLabel dynamic
Definition of all possible language codes in the schema is confusing and 
seems not a good way. Therefore we want to define it dynamically and 
therefore we have to allow selected dynamic properties. This is no 
problem as the logstash process (that we control) is the only process 
that writes data into elasticsearch

Ref https://gitlab.com/oersi/oersi-backend/-/issues/14